Add-PVR-Gallium-driver
The driver is essentially a Gallium frontend for the proprietary
Imagination Technologies driver. It makes use of the Gallium DRI
infrastructure, but little else.

Commit 9c772de270408c8a9b9e84b734d9fc0cee2d4265, "dri: Fold away
some unused indirection in __DriverAPIRec", has been reverted as
part of this change. The indirection removed by the commit is
needed by the PVR Gallium frontend, which does not use the same
context related code as dri2 and drisw.