egl-wayland-post-maximum-damage-when-blitting
When blitting, as part of the "is_different_gpu" case when swapping
buffers, the blit is done using the full surface dimensions, ignoring
the damage region. This results in corruption of the non damaged region
on screen. Workaround this by posting maximum damage when blitting.

A better fix would be to limit the blit to the damaged region, at least
when the number of damage rectangles is 1.